Add files via upload
This was the final release of GlovePIE that is rarely known.   The GloviePIE 0.46 was provided as a  standalone Executable on the developer's defunct Blog.  This version provided additional support for Wiimote PLUS and Kinect 360.   I combined the GlovePIE 0.46 Executable with the files provided by version 0.45.   GlovePIE is only compatible with the joystick emulator, PPjoy 0.8.4.6.  Enjoy.